requirements.txt
pyyaml==6.0.1
requests==2.31.0
pytest-bdd==7.0.1
pytest==8.0.0
assertpy==1.1
allure-pytest-bdd==2.13.2
pytest-rerunfailures==13.0